requirements.txt
Pillow
tqdm
playsound=1.2.2